backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.7.0
python-multipart>=0.0.9
structlog>=24.1.0
sqlalchemy>=2.0.30
asyncpg>=0.29.0
cryptography>=42.0.7
python-dotenv>=1.0.1
greenlet>=2.0.2